requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
jsonschema==4.17.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.9.1
pymongo==3.12.3
pyrsistent==0.19.2
pytz==2022.6
sqlparse==0.2.4
tomli==2.0.1
